requirements.txt
streamlit
python-dotenv
requests
uvicorn
fastapi
google-generativeai
python-multipart
